requirements.txt
customtkinter>=5.2.0
pillow>=11.1.0

